+++ NEWS.md
@@ -1,5 +1,7 @@
 # vctrs (development version)
 
+* `vec_equal()` now considers two objects with the same attribute names and values equivalent even if the attribute ordering is different. This reflects the fact that attributes are generally viewed as a map rather than an ordered list (#2156).
+
 * Pairlist tags and attribute names are now incorporated during object hashing.
   In practice this can come up when list elements are hashed inside functions
   like `vec_unique()` (#2154).